chore: Update ROUND_LIMIT constant to use usize instead of i32

+++ src/plugin/constants.rs
@@ -10,5 +10,5 @@ impl PluginConstants {
     pub const INITIAL_SALADS: i32 = 5;
     pub const INITIAL_CARROTS: i32 = 68;
 
-    pub const ROUND_LIMIT: i32 = 30;
+    pub const ROUND_LIMIT: usize = 30;
 }
